feat(cli): add shell completion support
Extract ~1100 lines of argparse parser construction from main() into
hermes_cli/parser.py as build_parser(). This keeps the parser tree
importable without triggering module-level side effects (.env loading,
heavy imports), which is required for shell completion via argcomplete.

- Add hermes completion zsh and hermes completion bash subcommands
  that output shell completion scripts (matching kubectl UX)
- Wire argcomplete.autocomplete(parser) in main() as a no-op unless
  the _ARGCOMPLETE env var is set by the shell
- Add PYTHON_ARGCOMPLETE_OK marker for global completion detection
- Add argcomplete>=3.0 as optional dependency (completion extra)
- Fix stale _SUBCOMMANDS set: add plugins, honcho, claw, acp, completion
- 10 inline cmd_* handlers moved as local functions in build_parser()
- 14 module-level handlers wired via lazy-import wrappers to avoid
  circular imports
- 8 new tests covering parser construction, completion output, graceful
  error handling, subcommand presence, and drift prevention

Closes NousResearch#2582

Made-with: Cursor
